Cloud2024/Cloud2024/src/python-lambdas/requirements.txt
python-ffmpeg-video-streaming
boto3
PyJWT>=2.3.0,<3.0.0